Rename to Inlet
