+++ src/axom/inlet/CMakeLists.txt
@@ -3,50 +3,48 @@
 #
 # SPDX-License-Identifier: (BSD-3-Clause)
 #------------------------------------------------------------------------------
-# SLIM component
+# Inlet component
 #------------------------------------------------------------------------------
 
 #------------------------------------------------------------------------------
 # Specify all headers/sources
 #------------------------------------------------------------------------------
-set(slim_headers
+set(inlet_headers
     Field.hpp
     GroupField.hpp
     IntField.hpp
     Map.hpp
     Structure.hpp
     Backend.hpp
-    MapBackend.hpp
-    )
+    MapBackend.hpp )
 
-set(slim_sources
-    Structure.cpp
-    )
+set(inlet_sources
+    Structure.cpp )
 
 
-blt_list_append(TO slim_headers ELEMENTS LuaMap.hpp IF LUA_FOUND)
-blt_list_append(TO slim_sources ELEMENTS LuaMap.cpp IF LUA_FOUND)
+blt_list_append(TO inlet_headers ELEMENTS LuaMap.hpp IF LUA_FOUND)
+blt_list_append(TO inlet_sources ELEMENTS LuaMap.cpp IF LUA_FOUND)
 
 #------------------------------------------------------------------------------
 # Make/Install the library
 #------------------------------------------------------------------------------
-set(slim_depends core slic fmt)
+set(inlet_depends core slic fmt)
 
-blt_list_append(TO slim_depends ELEMENTS lua IF LUA_FOUND)
+blt_list_append(TO inlet_depends ELEMENTS lua IF LUA_FOUND)
 
-blt_add_library( NAME        slim
-                 SOURCES     ${slim_sources}
-                 HEADERS     ${slim_headers}
-                 DEPENDS_ON  ${slim_depends}
-                 FOLDER      axom/slim
+blt_add_library( NAME        inlet
+                 SOURCES     ${inlet_sources}
+                 HEADERS     ${inlet_headers}
+                 DEPENDS_ON  ${inlet_depends}
+                 FOLDER      axom/inlet
                  OBJECT      TRUE
                  )
 
-axom_write_unified_header(NAME    slim
-                          HEADERS ${slim_headers})
+axom_write_unified_header(NAME    inlet
+                          HEADERS ${inlet_headers})
 
-axom_install_component(NAME      slim
-                       HEADERS   ${slim_headers}
+axom_install_component(NAME      inlet
+                       HEADERS   ${inlet_headers}
                        )
 
 
@@ -67,4 +65,4 @@ endif()
 #------------------------------------------------------------------------------
 # Add code checks
 #------------------------------------------------------------------------------
-axom_add_code_checks(PREFIX slim)
+axom_add_code_checks(PREFIX inlet)
