+++ src/axom/inlet/tests/CMakeLists.txt
@@ -3,7 +3,7 @@
 #
 # SPDX-License-Identifier: (BSD-3-Clause)
 #------------------------------------------------------------------------------
-# Tests for SLIM component
+# Tests for Inlet component
 #------------------------------------------------------------------------------
 
 #------------------------------------------------------------------------------
@@ -11,20 +11,17 @@
 #------------------------------------------------------------------------------
 
 # Add Serial GTests based tests
-set(gtest_slim_tests
-    slim_LuaMap.cpp
-    slim_Structure.cpp
-    )
+set(gtest_inlet_tests
+    inlet_LuaMap.cpp
+    inlet_Structure.cpp )
 
-foreach(test ${gtest_slim_tests})
+foreach(test ${gtest_inlet_tests})
     get_filename_component( test_name ${test} NAME_WE )
-    blt_add_executable( NAME ${test_name}_test
-                        SOURCES ${test}
+    blt_add_executable( NAME       ${test_name}_test
+                        SOURCES    ${test}
                         OUTPUT_DIR ${TEST_OUTPUT_DIRECTORY}
-                        DEPENDS_ON slim gtest 
-                        FOLDER axom/slim/tests )
+                        DEPENDS_ON inlet gtest 
+                        FOLDER axom/inlet/tests )
     blt_add_test( NAME    ${test_name} 
                   COMMAND ${test_name}_test )
 endforeach()
-
-
